+++ src/Language/JVM/Attribute.hs
@@ -15,11 +15,10 @@ This is the main module for accessing all kinds of Attributes.
 module Language.JVM.Attribute
   ( Attribute (..)
   , aInfo
-
   , aName
 
-  , IsAttribute (..)
   -- * SubAttributes
+  , IsAttribute (..)
   , Code
   , codeStackMapTable
   , ConstantValue
@@ -29,18 +28,22 @@ module Language.JVM.Attribute
 
   -- * Helpers
   , Const
+  , matching
+  , firstOne
   ) where
 
 import           GHC.Generics                            (Generic)
 
+import           Data.Foldable
+import           Data.List as List
 import           Data.Bifunctor
+import           Data.Maybe
 import           Data.Binary
 import qualified Data.ByteString                         as BS
 import qualified Data.ByteString.Lazy                    as BL
-import           Data.Monoid
 import           Data.Text                               as Text
 
-import           Language.JVM.Constant                   (ConstantPool, Index, deref)
+import           Language.JVM.Constant
 import           Language.JVM.Utils                      (SizedByteString32,
                                                           trd,
                                                           unSizedByteString)
@@ -67,8 +70,8 @@ aInfo = unSizedByteString . aInfo'
 
 -- | Extracts the name from the attribute, if it exists in the
 -- ConstantPool.
-aName :: ConstantPool -> Attribute -> Maybe Text.Text
-aName cp = deref cp . aNameIndex
+aName :: Attribute -> PoolAccess Text.Text
+aName = derefF aNameIndex
 
 -- | Create a type dependent on another type 'b',
 -- used for accessing the correct 'attrName' in 'IsAttribute'.
@@ -85,21 +88,30 @@ class IsAttribute a where
   -- ^ Generate a 'a' from an Attribute.
 
   fromAttribute'
-    :: ConstantPool
-    -> Attribute
-    -> Maybe (Either String a)
-  fromAttribute' cp as = do
-    name <- aName cp as
-    if name == unConst (attrName :: Const Text.Text a) then
-       return $ fromAttribute as
-    else Nothing
+    :: Attribute
+    -> PoolAccess (Maybe (Either String a))
+  fromAttribute' as = do
+    name <- aName as
+    return $ if name == unConst (attrName :: Const Text.Text a)
+      then Just $ fromAttribute as
+      else Nothing
+
+-- | Return a list of either parsed attributes or error messages from an object
+-- given a function that object to a collection of attributes. The list only
+-- contains matching attributes.
+matching
+  :: (IsAttribute a, Traversable t)
+  => (b -> t Attribute)
+  -> b
+  -> PoolAccess [Either String a]
+matching fn b = do
+  catMaybes . toList <$> sequence (fromAttribute' <$> fn b)
 
 readFromStrict :: Binary a => Attribute -> Either String a
 readFromStrict =
     bimap trd trd . decodeOrFail . BL.fromStrict . aInfo
 
 -- # Attributes
-
 
 -- | 'Code' is an Attribute.
 instance IsAttribute Code where
@@ -130,9 +142,13 @@ instance IsAttribute BootstrapMethods where
 -- 'Attribute'. This is a small hack to fix it.
 type Code = Code.Code Attribute
 
+-- | Maybe return the first element of a list
+firstOne :: [a] -> Maybe a
+firstOne as = fst <$> List.uncons as
+
 -- | The a attribute of the code is the StackMapTable. There can be at most one.
 -- If the version number of the file is more than 50, and there is no StackMapTable.
 -- there is an implicit empty StackMapTable.
-codeStackMapTable :: ConstantPool -> Code -> Maybe (Either String BootstrapMethods)
-codeStackMapTable cp =
-  getFirst . foldMap (First . fromAttribute' cp) . Code.attributes
+codeStackMapTable :: Code -> PoolAccess (Maybe (Either String BootstrapMethods))
+codeStackMapTable c = do
+  firstOne <$> matching Code.attributes c