+++ src/Language/JVM/Constant.hs
@@ -1,3 +1,6 @@
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE DeriveFunctor #-}
+{-# LANGUAGE EmptyCase #-}
 {-|
 Module      : Language.JVM.Constant
 Copyright   : (c) Christian Gram Kalhauge, 2017
@@ -32,8 +35,12 @@ module Language.JVM.Constant
   , ConstantPool (..)
 
   , Index(..)
-  , indexAsWord
   , InConstantPool (..)
+
+  , PoolAccess(..)
+  , PoolAccessError(..)
+
+  , derefF
 
   ) where
 
@@ -51,7 +58,9 @@ import           Language.JVM.Utils
 import           Language.JVM.Type
 
 import qualified Data.Text          as Text
+import qualified Data.ByteString    as BS
 import qualified Data.Text.Encoding as TE
+import qualified Data.Text.Encoding.Error as TE
 
 -- | A constant is a multi word item in the 'ConstantPool'. Each of
 -- the constructors are pretty much self expiatory from the types.
@@ -71,6 +80,9 @@ data Constant
   | MethodType (Index MethodDescriptor)
   | InvokeDynamic !Word16 (Index Constant)
   deriving (Show, Eq)
+
+typeToStr :: Constant -> String
+typeToStr = head . words . show
 
 instance Binary Constant where
   get = do
@@ -180,19 +192,35 @@ derefConstant :: ConstantPool -> Word16 -> Maybe Constant
 derefConstant (ConstantPool cp) ref =
   IM.lookup (fromIntegral ref) cp
 
+-- | A pool access error
+data PoolAccessError
+  = PoolAccessError Word16 String
+  deriving (Show, Eq)
+
+-- | The pool access monad.
+newtype PoolAccess a = PoolAccess
+  { runWithPool :: ConstantPool -> Either PoolAccessError a
+  } deriving (Functor)
+
+instance Applicative (PoolAccess) where
+  pure x = PoolAccess . const $ Right x
+  mf <*> ma = PoolAccess $ \c ->
+     runWithPool mf c <*> runWithPool ma c
+
+instance Monad (PoolAccess) where
+  return = pure
+  ma >>= fm = PoolAccess $ \c ->
+    case runWithPool ma c of
+      Left p -> Left p
+      Right b -> runWithPool (fm b) c
+
 -- This part contains helpers to describe some of the semantics of the
 -- class file.
 
 -- | Describes an index into the constant pool
-data Index x where
-  Index :: (InConstantPool x) => Word16 -> Index x
-
--- | Get the internal word of the index
-indexAsWord :: Index x -> Word16
-indexAsWord (Index x) = x
-
-deriving instance (Show (Index x))
-deriving instance (Eq (Index x))
+newtype Index x = Index
+  { indexAsWord :: Word16
+  } deriving (Show, Eq)
 
 instance (InConstantPool x) => Binary (Index x) where
   get = Index <$> get
@@ -200,52 +228,109 @@ instance (InConstantPool x) => Binary (Index x) where
 
 -- | Describes if a type is in the constant pool
 class InConstantPool a where
-  deref :: ConstantPool -> Index a -> Maybe a
+  deref :: Index a -> PoolAccess a
 
 instance InConstantPool Constant where
-  deref cp (Index ref) =
-    derefConstant cp ref
+  deref (Index ref) = PoolAccess $ \cp ->
+    case derefConstant cp ref of
+      Just c -> Right c
+      Nothing -> runWithPool (outofbounds ref) cp
 
 -- | Lookup a 'Text.Text' in the 'ConstantPool', returns 'Nothing' if the
 -- reference does not point to something in the ConstantPool, if it points to
 -- something not a 'Language.JVM.Constant.String' Constant, or if it is
 -- impossible to decode the 'ByteString' as Utf8.
 instance InConstantPool Text.Text where
-  deref cp (Index ref) = do
-    String str <- derefConstant cp ref
-    case TE.decodeUtf8' . unSizedByteString $ str of
-      Left _    -> Nothing
-      Right txt -> Just txt
+  deref (Index ref) = do
+    cons <- derefW ref
+    case cons of
+      String str ->
+        case TE.decodeUtf8' . unSizedByteString $ str of
+          Left (TE.DecodeError msg _) -> badEncoding ref msg (unSizedByteString str)
+          Left _ -> error "This is deprecated in the api"
+          Right txt -> return txt
+      a -> wrongType ref "String" a
 
 instance InConstantPool ClassName where
-  deref cp (Index ref) = do
-    ClassRef r <- derefConstant cp ref
-    ClassName <$> deref cp r
+  deref (Index ref) = do
+    cons <- derefW ref
+    case cons of
+      ClassRef r ->
+        ClassName <$> deref r
+      a -> wrongType ref "ClassRef" a
 
 instance InConstantPool (InClass MethodId) where
-  deref cp (Index ref) = do
-    MethodRef cn rt <- derefConstant cp ref
-    InClass <$> deref cp cn <*> deref cp rt
+  deref (Index ref) = do
+    cons <- derefW ref
+    case cons of
+      MethodRef cn rt ->
+        InClass <$> deref cn <*> deref rt
+      a -> wrongType ref "MethodRef" a
 
 instance InConstantPool (InClass FieldId) where
-  deref cp (Index ref) = do
-    FieldRef cn rt <- derefConstant cp ref
-    InClass <$> deref cp cn <*> deref cp rt
+  deref (Index ref) = do
+    cons <- derefW ref
+    case cons of
+      FieldRef cn rt ->
+        InClass <$> deref cn <*> deref rt
+      a -> wrongType ref "FieldRef" a
 
 instance InConstantPool MethodId where
-  deref cp (Index ref) = do
-    NameAndType n (Index t) <- derefConstant cp ref
-    MethodId <$> deref cp n <*> deref cp (Index t)
+  deref (Index ref) = do
+    cons <- derefW ref
+    case cons of
+      NameAndType n (Index i) ->
+        MethodId <$> deref n <*> derefW i
+      a -> wrongType ref "NameAndType" a
 
 instance InConstantPool FieldId where
-  deref cp (Index ref) = do
-    NameAndType n (Index t) <- derefConstant cp ref
-    FieldId <$> deref cp n <*> deref cp (Index t)
+  deref (Index ref) = do
+    cons <- derefW ref
+    case cons of
+      NameAndType n (Index i) ->
+        FieldId <$> deref n <*> derefW i
+      a -> wrongType ref "NameAndType" a
 
 instance InConstantPool MethodDescriptor where
-  deref cp (Index ref) = do
-    methodDescriptorFromText =<< deref cp (Index ref)
+  deref (Index ref) = do
+    txt <- derefW ref
+    case methodDescriptorFromText txt of
+      Just x -> return x
+      Nothing -> parseError ref txt
 
 instance InConstantPool FieldDescriptor where
-  deref cp (Index ref) = do
-    fieldDescriptorFromText =<< deref cp (Index ref)
+  deref (Index ref) = do
+    txt <- derefW ref
+    case fieldDescriptorFromText txt of
+      Just x -> return x
+      Nothing -> parseError ref txt
+
+-- Helper functions
+
+-- | Given a function from an object b to an 'Index a' create a function from b
+-- to 'PoolAccess a'. Very useful for defining functions.
+derefF :: InConstantPool a => (b -> Index a) -> b -> PoolAccess a
+derefF fn b = deref (fn b)
+
+-- Hidden access method.
+
+derefW :: InConstantPool a => Word16 -> PoolAccess a
+derefW w = deref (Index w)
+
+-- Hidden exception methods
+
+wrongType :: Word16 -> String -> Constant -> PoolAccess a
+wrongType w expected got = PoolAccess $ \_ ->
+  Left (PoolAccessError w ("Expected '" ++ expected ++ "' but got '" ++ typeToStr got ++ "'"))
+
+outofbounds :: Word16 -> PoolAccess a
+outofbounds w = PoolAccess $ \_ ->
+  Left (PoolAccessError w "Out of bounds")
+
+parseError :: Word16 -> Text.Text -> PoolAccess a
+parseError w t = PoolAccess $ \_ ->
+  Left (PoolAccessError w $ "Could not parse '" ++ Text.unpack t ++ "'")
+
+badEncoding :: Word16 -> String -> BS.ByteString -> PoolAccess a
+badEncoding w str bs = PoolAccess $ \_ ->
+  Left (PoolAccessError w $ "Could not encode '" ++ str ++ "': " ++ show bs)